feat(Who-command): the Who command now correctly hides users who haven't posted publicly in the last 14 days
the Who command lists all users who are online, but in order to honor users' privacy, will not list users (except to sysops and aides) who haven't posted publicly, who are stealth.  this is intended to reduce doxxing, particularly on boards which deal with sensitive topics where users may not wish to be known.  sysops and aides remain able to see everyone who's online, for system security reasons.